FTR: share filesystem paths across EWMRS and WPC
