+++ src/common/ingest/wpc/downloader.py
@@ -8,7 +8,7 @@ from typing import Optional, Tuple
 from pathlib import Path
 
 from common.ingest.wpc.config import WPC_CODED_SFC_BASE_URL, VALID_HOURS
-from util.ewmrs_file import WPC_SFC_DIR
+from util.file import WPC_SFC_DIR
 from util.io import IOManager
 
 io_manager = IOManager("[WPC]")
